CIRC-717: Implement logic to check if allowed refund interval has not exceeded.

+++ src/test/java/api/support/fixtures/TemplateContextMatchers.java
@@ -230,7 +230,7 @@ public class TemplateContextMatchers {
       hasJsonPath("fee.actionType", is(action.getTypeAction())),
       hasJsonPath("fee.actionDateTime", isEquivalentTo(action.getDateAction())),
       hasJsonPath("fee.actionAmount", is(action.getAmountAction().doubleValue())),
-      hasJsonPath("fee.amount", is(account.getAmount().doubleValue())),
+      hasJsonPath("fee.amount", is(account.getAmount())),
       hasJsonPath("fee.balance", is(action.getBalance()))
     );
   }
